Phase 2: Migrate SettingsContext to all pages and extract useNotifications hook
Track A — SettingsContext migration to all pages and child components:

Propagated useSettings() context consumption to all 7 remaining pages and
6 child components, eliminating ~65 prop-drilled settings assignments from
App.jsx's route JSX blocks.

Pages updated (settings props removed → now consumed from context):
- AnnouncePage: removed 24 settings props (74→50 total props)
- PlayByPlayPage: removed 15 settings props (65→50)
- EmceePage: removed 12 settings props (26→14)
- AllianceSelectionPage: removed 6 settings props (34→28)
- TeamDataPage: removed 4 settings props (29→25)
- SchedulePage: removed 2 settings props (29→27)
- RanksPage: removed 2 settings props (25→23)

Child components now consuming useSettings() directly:
- Announce.jsx: 11 settings (was forwarded from AnnouncePage)
- PlayByPlay.jsx: 4 settings (was forwarded from PlayByPlayPage)
- TopButtons.jsx: swapScreen (timeFormat removed, MatchClock gets its own)
- BottomButtons.jsx: highScoreMode, showWorldAndStatsOnAnnouncePlayByPlay
- MatchClock.jsx: timeFormat
- EmceeClock.jsx: timeFormat

Added useFourTeamAlliances to SettingsContext (was a usePersistentState in
App.jsx, now lives alongside the other 28 display/UX settings).

Track B — useNotifications hook extraction:

Created src/hooks/useNotifications.js (351 lines) extracting all notification
system code from App.jsx:

State owned by hook (5 usePersistentState variables):
- systemMessage, systemBell — system-wide notifications
- eventMessage, eventBell — per-event notifications
- showReloaded — app update notification flag

API functions moved to hook:
- putNotifications, putEventNotifications — write notification config
- getNotifications, getEventNotifications — fetch raw notification data
- getSystemMessages, getEventMessages — fetch + format + set state

Effects moved to hook:
- Service worker update available snackbar (showReload + waitingWorker)
- Post-reload update details snackbar (showReloaded + appUpdates)

Hook internally calls useOnlineStatus(), useServiceWorker(), and
useSnackbar() — no need to pass these from App.jsx. Parameters are:
httpClient, selectedEvent, useFTCOffline, manualOfflineMode.

App.jsx still calls getSystemMessages/getEventMessages in its own auto-refresh
and loadEvent effects via destructured references from the hook.

Cleanup:
- Removed unused Button import from App.jsx (was only in snackbar JSX)
- Removed unused useSwipe from App.jsx useSettings() destructure
- Removed unused getEventNotifications from App.jsx hook destructure

Build: Compiled successfully with zero warnings.
App.jsx: 7,550 → 7,310 lines (~240 lines removed)
Total prop assignments eliminated (Phase 1+2): ~121
