+++ src/components/EmceeClock.jsx
@@ -1,8 +1,10 @@
 import { useState, useEffect } from 'react';
 import moment from 'moment/moment';
 import { useInterval } from 'react-interval-hook';
+import { useSettings } from "../contexts/SettingsContext";
 
-const EmceeClock = ({ matchDetails, timeFormat }) => {
+const EmceeClock = ({ matchDetails }) => {
+    const { timeFormat } = useSettings();
     const [currentTime, setCurrentTime] = useState(moment());
 
     const { start, stop } = useInterval(
