+++ src/pages/TeamDataPage.jsx
@@ -19,6 +19,7 @@ import 'react-quill/dist/quill.snow.css';
 import TeamTimer from "components/TeamTimer";
 import { useInterval } from "react-interval-hook";
 import useScrollPosition from "../hooks/useScrollPosition";
+import { useSettings } from "../contexts/SettingsContext";
 
 /** Blue banner stats are a nested object; SheetJS leaves those cells blank unless stringified. */
 const BLUE_BANNER_EXPORT_ROWS = [
@@ -66,7 +67,8 @@ function formatBlueBannersForExport(bb) {
     return parts.join("; ");
 }
 
-function TeamDataPage({ selectedEvent, selectedYear, teamList, rankings, teamSort, setTeamSort, communityUpdates, setCommunityUpdates, allianceCount, lastVisit, setLastVisit, putTeamData, localUpdates, setLocalUpdates, qualSchedule, playoffSchedule, originalAndSustaining, monthsWarning, user, isAuthenticated, getTeamHistory, timeFormat, getCommunityUpdates, getTeamList, eventLabel, ftcMode, remapNumberToString, useScrollMemory }) {
+function TeamDataPage({ selectedEvent, selectedYear, teamList, rankings, teamSort, setTeamSort, communityUpdates, setCommunityUpdates, allianceCount, lastVisit, setLastVisit, putTeamData, localUpdates, setLocalUpdates, qualSchedule, playoffSchedule, originalAndSustaining, user, isAuthenticated, getTeamHistory, getCommunityUpdates, getTeamList, eventLabel, ftcMode, remapNumberToString }) {
+    const { monthsWarning, timeFormat, useScrollMemory } = useSettings();
     const [currentTime, setCurrentTime] = useState(moment());
     const [clockRunning, setClockRunning] = useState(true);
     const { disableScope, enableScope } = useHotkeysContext();
